// WADNR.Web/src/app/shared/generated/model/project-funding-save-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectFundSourceAllocationRequestItemRequest } from './project-fund-source-allocation-request-item-request';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectFundingSaveRequest { 
    EstimatedTotalCost?: number | null;
    FundingSourceNotes?: string | null;
    FundingSourceIDs?: Array<number> | null;
    AllocationRequests?: Array<ProjectFundSourceAllocationRequestItemRequest> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectFundingSaveRequestForm { 
    EstimatedTotalCost?: FormControl<number>;
    FundingSourceNotes?: FormControl<string>;
    FundingSourceIDs?: FormControl<Array<number>>;
    AllocationRequests?: FormControl<Array<ProjectFundSourceAllocationRequestItemRequest>>;
}

export class ProjectFundingSaveRequestFormControls { 
    public static EstimatedTotalCost = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FundingSourceNotes = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FundingSourceIDs = (value: FormControlState<Array<number>> | Array<number> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<number>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static AllocationRequests = (value: FormControlState<Array<ProjectFundSourceAllocationRequestItemRequest>> | Array<ProjectFundSourceAllocationRequestItemRequest> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectFundSourceAllocationRequestItemRequest>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}